Keep clone pagination counters in 64 bits
rdf_fdw_clone_table() walks a result in pages, advancing an offset by the page
size after each one and accumulating the number of rows it has inserted. Both
were held in an int.

An offset that passes 2147483647 therefore wraps. Starting at a begin_offset
of 2000000000 with a page of 1000000000, the first request is correct and the
second goes out as

  OFFSET -1294967296 LIMIT 1000000000

and the ones after it as -294967296, then 705032704, wandering over the result
rather than walking it: some rows are read twice and others are never reached.
It is the same wrap that OFFSET pushdown had before it was read as an int64.

Hold the offset and the row count in int64, compute the remaining page size by
subtraction rather than by adding and comparing, and check each addition
before making it. Carry the width through the request text and the progress
line.

The offset argument of the page request also loses a ternary that sent 0 on
the first page when begin_offset was 0. state->offset is initialised to
begin_offset, so it already held 0 in exactly that case.

Only the C files of the original change are taken. Its third file adds a rule
to scripts/review/defensive-source.py, which belongs to the review harness the
series was produced with and not to this repository.

No test comes with this. Showing the wrap needs a clone to finish one page and
begin another past the int range, which means an endpoint that answers
whatever offset it is given -- the stub does, and that is how it was shown --
but such an endpoint never returns the empty page that ends the loop, and
bounding the loop with max_records caps the page size below what overflows.

+++ rdf_fdw.h
@@ -249,9 +249,9 @@ typedef struct RDFfdwState
 	char *target_table_name;
 	char *ordering_pgcolumn;
 	char *sort_order;
-	int offset;
+	int64 offset;
 	int fetch_size;
-	int inserted_records;
+	int64 inserted_records;
 	int batch_size; 				   /* Number of rows to batch for INSERT/UPDATE/DELETE */
 	int batch_count;				   /* Current number of rows in the batch buffer */
 	StringInfoData batch_statements;   /* Buffer for batched SPARQL statements */
